Split off `TagsView` + sorted out proper colors
