+++ Spawn-App-iOS-SwiftUI/Views/FriendsView.swift
@@ -21,10 +21,7 @@ struct FriendsView: View {
             VStack(spacing: 20) {
                 header
                 searchBar
-                tagSection
-                closeFriendsSection
-                Spacer()
-                otherTagsSection
+                TagsView(user: user)
             }
             .padding()
             .background(universalBackgroundColor)
@@ -67,71 +64,6 @@ private extension FriendsView {
                 .padding(.horizontal, 10)
         }
         .frame(height: 40)
-    }
-    
-    var tagSection: some View {
-        VStack(alignment: .leading, spacing: 15) {
-            Text("TAGS")
-                .font(.headline)
-            
-            AddTagButton()
-        }
-        .padding(.top)
-    }
-    
-    var closeFriendsSection: some View {
-        VStack(spacing: 15) {
-            HStack {
-                Text("Close Friends")
-                    .font(.headline)
-                Spacer()
-                EditButton()
-            }
-            
-            colorOptions
-            
-            VStack(spacing: 10) {
-                if let friends = user.friends, !(user.friends?.isEmpty ?? false) {
-                    ForEach(friends) { friend in
-                        FriendRow(friend: friend)
-                    }
-                }
-            }
-        }
-        .padding()
-        .background(RoundedRectangle(cornerRadius: 12).fill(Color.blue.opacity(0.2)))
-    }
-    
-    var colorOptions: some View {
-        HStack(spacing: 15) {
-            ForEach(0..<5) { index in
-                Circle()
-                    .fill(index == 4 ? Color.gray.opacity(0.2) : randomColor(index))
-                    .frame(width: 30, height: 30)
-                    .overlay(
-                        index == 4
-                        ? Image(systemName: "plus").foregroundColor(.black)
-                        : nil
-                    )
-            }
-        }
-    }
-    
-    var otherTagsSection: some View {
-        ScrollView {
-            VStack(spacing: 15) {
-                ForEach(0..<3) { index in
-                    TagRow(tagName: "Tag Name", color: randomColor(index))
-                }
-            }
-        }
-    }
-}
-
-private extension FriendsView {
-    func randomColor(_ index: Int) -> Color {
-        let colors: [Color] = [.purple, .green, .red, .blue, .gray]
-        return colors[index % colors.count]
     }
 }
